fixed gcc warning

+++ Renderer/GL/GLVolume2DTex.cpp
@@ -96,7 +96,7 @@ void GLVolume2DTex::Bind(UINT32 iUnit,
                          int iDepth,
                          int iStack) {
   
-  if (iDepth > 0 && iDepth < m_pTextures[iStack].size()) {
+  if (iDepth > 0 && iDepth < static_cast<int>(m_pTextures[iStack].size())) {
     m_pTextures[iStack][iDepth]->Bind(iUnit);
   } else {
     // mirrored clamp
